fixed the unstable indirect problem. Moved logic from data provider to the TemplatedListResultTable
